feat: add UI mode store and error boundary
