+++ src/hooks/useGameClock.ts
@@ -1,9 +1,10 @@
 import { useEffect } from "react";
 import { useGameStore } from "@/state/gameStore";
+import { useUIStore } from "@/state/ui";
 
 export function useGameClock() {
   const paused = useGameStore((s) => s.ui.paused);
-  const mode = useGameStore((s) => s.ui.mode);
+  const mode = useUIStore((s) => s.mode);
   const tick = useGameStore((s) => s.tick);
 
   useEffect(() => {
